Use Ui for interactions with the user
